refactor: room card into a component
